+++ src/Catty/PlaceAtBrick.m
@@ -42,7 +42,7 @@
     NSDebug(@"Performing: %@", self.description);
     double xPosition = [self.xPosition interpretDoubleForSprite:self.object];
     double yPosition = [self.yPosition interpretDoubleForSprite:self.object];
-    self.object.position = CGPointMake(xPosition, yPosition);
+    self.object.position = CGPointMake((CGFloat)xPosition, (CGFloat)yPosition);
   };
 }
