+++ src/Catty/SetYBrick.m
@@ -42,8 +42,8 @@
 {
   return ^{
     NSDebug(@"Performing: %@", self.description);
-    float yPosition = [self.yPosition interpretDoubleForSprite:self.object];
-    self.object.position = CGPointMake(self.object.xPosition, yPosition);
+    double yPosition = [self.yPosition interpretDoubleForSprite:self.object];
+    self.object.position = CGPointMake(self.object.xPosition, (CGFloat)yPosition);
   };
 }
